Optimizing ab_pruning. Work in Progress
Trying to optimize the ab_pruning so that it does not need to run miniMax when a forced move is needed.  This should reduce some of the overhead.

There is a problem right now where the forcedMoves need to be called from miniMax but I'm uncertain how / what to be returning.